+++ libs/base/src/num/niche/mem/mod.rs
@@ -1,11 +1,10 @@
-// devela::num::niche::mem
+// devela_base::num::niche::mem
 //
 //! Numeric types for memory-efficient representations.
 //
 
 #[cfg(test)]
 mod tests;
-// mod asm_validation;
 
 mod non_value;
 pub use non_value::*;