+++ libs/base/src/num/niche/mem/tests.rs
@@ -1,4 +1,4 @@
-// devela::num::niche::tests
+// devela_base::num::niche::mem::tests
 //
 //!
 //